# AK+Userland: Introduce custom error payloads and start using them for Streams