--- a/AK/Forward.h
+++ b/AK/Forward.h
@@ -30,6 +30,8 @@ class DeprecatedStringCodePointIterator;
 class Duration;
 class Error;
 class FlyString;
+class FormatBuilder;
+struct FormatString;
 class GenericLexer;
 class IPv4Address;
 class JsonArray;
@@ -105,6 +107,12 @@ class FixedArray;
 template<size_t precision, typename Underlying = i32>
 class FixedPoint;
 
+template<typename T, typename = void>
+struct Formatter;
+
+template<>
+struct Formatter<FormatString>;
+
 template<typename>
 class Function;
 
@@ -171,6 +179,9 @@ using AK::ErrorOr;
 using AK::FixedArray;
 using AK::FixedPoint;
 using AK::FlyString;
+using AK::FormatBuilder;
+using AK::FormatString;
+using AK::Formatter;
 using AK::Function;
 using AK::GenericLexer;
 using AK::HashMap;