For various file formats, we need to know whether we failed to read a value due to encountering EOF or if it errored due to a different reasons.

Since comparing string literals is not great and we can't easily add a new errno due to portability problems with Lagom, this adds a new, rather generic storage area where users can basically store arbitrary error data. The different types are discerned via vtables and `dynamic_cast`, so we should only have a performance impact (beyond the impact caused by the slightly increased size) whenever we interact with the error payload itself, not when interacting with the surrounding `Error`.

The performance values look like they are reasonably within margin of error, at least to a point where we don't have a measurable performance decrease:

|                                                  | Before  | After   |
|--------------------------------------------------|---------|---------|
| `tar -tf llvm-project-15.0.3.src.tar.xz` (n=10)  | 18920ms | 18275ms |
| `/usr/Tests/LibCompress/TestXz` (n=1000)         | 42ms |  43ms  |
| `test-js` (n=100)                                | 5241ms | 5080ms |

|                                                  | Before  | After   | Difference | Deviation(?) |
|--------------------------------------------------|---------|---------|---------|---------|
| `./bin/tar -tf llvm-project-15.0.3.src.tar.xz` (n=1)         | 153056076473 instructions | 162921936369 instructions | +6.4% | - |
| `./bin/TestXz` (n=5)         |  9211626 instructions | 9524613 instructions | +3.4% | $10^2$ |
| `./bin/test-js` (n=5)                                | 30863526661 instructions | 32203265126 instructions | +4.3% | $10^8$ |

Remaining points of discussion (maybe):
 - Naming of the new classes
 - Have I implemented the formatters correctly
 - ~~What to do in Kernel land~~