@timschumi 2023-09-27
Boop. I forgot what I was planning to do. Probably getting some more performance metrics.